requirements.txt
torch>=2.0.0
pandas
numpy
scikit-learn
pyyaml
networkx
scapy
termcolor
matplotlib
seaborn
